Maintain and display constraints history component (#47)
I implemented the constraints Current section. The History section is there, but will not get updated yet. Both sections also will show an empty state component to inform the user. Tests have been added for the state logic of adding and deleting constraints. But only deleting has been tested for the UI. This is because the component does not offer a way to add constraints. They must be done through the constraint components.

Squashed commits:
* Add close button

* Display non ideal state when no constraints are set

* Display current constraints in query controller

* Display history of constraints

* Divide the current and history sections

* Delete constraints when remove icon is clicked

* Add constraints to current query

* Add stories for view all pop up

* Test: delete a current constraint in QueryControllerMachine

* Test: displays empty state when no constraints are set

* Decrease padding between the title and constraints list

* Add yarn unplugged to gitignore