Build the UI for a suggest box constraint (#51)
I created a reusable constraint component that handles constraints with suggest boxes, e.g. pathway name or GO Annotation. Then when a user begins typing, a weighted search is performed across the available values, with the best choice as the first. The user can then select the choice by pressing enter, or clicking on a relevant result which places it in the selected constraint lists for this component. The user can delete it from there.


Squashed commits:
* Display NonIdealState when no constraint values are present

* Display a story for constraints not being set

* Display story for when constraints change

* Render only constraints that are not set

* Display story for when constraints are applied

* Remove an item from the list of selected constraints

* Display story for when constraints are not set

* Fuzzy search the values for matches

* Reset the input value on select event

* Inform the user if their search returns nothing