Display applied constraints in the query controller (#55)
The default organisms are now being rendered into the main UI, and are connected to the constraint query controller. The constraints are then synced between the controller and the related individual constraints.

I had to do a quick redesign in order to show and manage the constraints better. The selected values are now being displayed under a constraint name header.

Squashed commits:
* Render the correct constraint buttons

* Allow building project on tsc errors

* Render constraint UIs using a user config

* Disable create-react-app typechecking

* Dispatch a global query object to the controller

* Render applied constraints

* Display the default constraints

* Apply constraints to the next query

* Display applied constraints as a nested list

* Display applied constraint values with their count

* Allow removing constraints from the query controller

* Reset deleted constraint when deleted from Query Controller

* Make popup card wider