Display how many constraints are applied (#64)
Constraint tags were not being displayed. Now if a constraint has been
updated, it will display a tag (with a number if multiple exist).

If the tag is applied to the next query, it will be green. Otherwise it
will be yellow to indicate that the constraint has items present but not
applied.

Squashed commits:

* Show if constraint items have been applied or updated