Paginate the table row data (#100)
To prevent excessive network calls, this commit caches the retrieved
table rows into a `Map` of pages. It will fetch the total cache factor,
which is the amount of pages a user wishes to retrieve per request.

The cache is not currently limited to a max size, and is busted with
every new query.

The user now has 3 options for navigating the table forwards or backwards.

1. next/previous page
2. first/last page
3. direct page number input

The page navigation click events are debounced. This prevents having the user
wait on the results of fetching new results when they reach a page that
is not cached.

Closes: #99

Squashed commits:
* Display row count information
* Start a cache from the initial table row results
* Navigate forward if page exists in the cache
* Navigate backwards if page exists in the cache
* Fetch next set of rows when navigating forward and page not cached
* Reset the page number when updating the query
* Enable changing pages by inputting a number
* Display total number of pages
* Add buttons to skip to 1st and last page
* Fetch next set of rows when navigating backward and page not cached
* Bust the cache when fetching new queries
* Debounce successive page navigation click events