Perf: Use xstate's assign rather than immer (#109)
`@xstate/immer` provides a nicer DX for updating the context. But it
fails miserably with large datasets. This is because it needs to
recursively convert each item into an immutable draft. Considering that
intermine produces enormous amounts of data, and we don't modify that
data, we will not have an issue with mutations there.

The only place where mutation would cause problems is with our data
structures for state management. This is alleviated by using `xstate`'s
assign method.

Therefore, for performance reasons, this commit refactors to using
`xstate`s assign.

Closes: #108

Squashed commits:
* Refactor appManagerMachine
* Refactor barChartMachine
* Refactor overviewConstraintMachine
* Refactor pieChartMachine
* Refactor queryControllerMachine
* Refactor tableMachine
* Refactor templateConstraintMachine
* Refactor templateQueryMachine