Persist machines between component mounts (#113)
`xstate` machines are stopped whenever the component mounts if we use
its `useMachineBus` method. This is not ideal, because it means that
the components have to refetch their values on every fresh mount. Taking
into account that some mines may have large latency, we need to keep the
api calls running in the background. The idiomatic way to prevent this,
would be to create services and pass that to the child components.

The issue with that is that we have nested components with child nested
components, that each make their own specific calls. Spawning new actors
using `xstate` would lead to hard to track down state bugs.

Therefore this commit provides a modified version of `useMachine` that
does not stop the running services. Instead, it will reuse one with a
similar machine id. This allows components to finish their api calls in
the background should a user click out of a popup panel.

Closes: #112

Squashed Commits:
* Remove serviceBusWrapper
* Refactor useMachineBus to enable persistence
* Persist constraint machines
* Persist PieChart machine
* Persist BarChart machine
* Persist Table machine
* Persist Template constraint machines
* Reload overview constraints on mine/class changes
* Cache templates when changing views
* Persist Query controller machine