Maintain the query controller state between constraint views (#143)
The overview queries could not be executed after switching to and back
from the template view. This occurred because the overview constraints
and the template view constraints were being spawned, and therefore not
initialized in the event bus as before. This commit also spawns the
query controller so that it can stay persistent like the other actors.

This commit also refactors the event bus to use hooks so the components
can explicitly register to the bus.

Closes: #142

Squashed commits:
* Spawn the query controller and register on the event bus
* Register components to the event bus
* Simplify event bus logic