Fix: checkmark tick not showing in checkboxes (#152)
Blueprintjs has a very particular sass setup, and it needs to be recreated
in order to fully work with their build. Since we are modifying the sass
variables, we are not able to replicate it, and therefore the inline
svgs are not being displayed.

This commit manually adds the checkmark inline ticks to the main sass
file. It was lifted from this comment in the Blueprintjs issues tracker
https://github.com/palantir/blueprint/issues/2821#issuecomment-414645690

Closes: #151