Generate code snippets for the current table results (#160)
This commit adds the feature for generating code snippets for the current
table results and query. Users will be able to set a file name if they
wish. Any fetched code will be cached locally, and will be busted if the
query changes. A text notification is displayed in the code section if
the code has yet to be loaded.

Closes: #159

Squashed commits:
* Display a dialogue for generating code snippets
* Fetch code on new language selection
* Highlight code snippets
* Cache fetched code for current query
* Display a notification if fetching code is in progress
* Save code file