Fix: the charts not exiting error states (#193)
When a chart was in an error state, for example when it doesn't receive
any values, it will stay there and not transition on new queries. This
occurred because the table machine would update the available selected
paths on each new query. And if it errors, then the paths would be reset
to an empty array.

This commit ensures that the paths are only updated when the mine changes,
and that the machines transition out of error states when the class or mine
changes as well.

Closes: #192