Do not display overview constraints that do not exist on class (#199)
Some classes do not have constraints available, and we can tell which
they are. Yet right now we still load them and display a warning to the
user. But this isn't ideal since it is a known state, and not an error.

This commit will only render those constraints that are available for the
chosen class. When no constraints exist for the chosen class, a
notification will be displayed to the user.

Closes: #198